feat: implement shared authentication instance and AsyncApexJob methods in ToolingClient
